+++ CHANGELOG.md
@@ -20,6 +20,7 @@
   [env]
   file = ".env"
   ```
+- Updated `proto upgrade` to error if there's another process of proto currently running.
 
 #### 🐞 Fixes